% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation_system.R
\name{add_deprecation_warning}
\alias{add_deprecation_warning}
\title{Function Deprecation System}
\usage{
add_deprecation_warning(
  function_name,
  replacement_function = NULL,
  removal_version = "2.0.0"
)
}
\arguments{
\item{function_name}{Name of function to deprecate}

\item{replacement_function}{Suggested replacement function}

\item{removal_version}{Version when function will be removed}
}
\value{
Deprecation warning message
}
\description{
Implements function deprecation with warnings and migration guides
}
\keyword{internal}
