% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-export.R
\name{add_summary_charts}
\alias{add_summary_charts}
\title{Add Summary Charts to Excel Workbook}
\usage{
add_summary_charts(wb, summary_data)
}
\description{
Add Summary Charts to Excel Workbook
}
\keyword{internal}
