% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching_internal.R
\name{aliases_long}
\alias{aliases_long}
\title{Convert aliases list column to long form (internal)}
\usage{
aliases_long(roster_df)
}
\description{
Convert aliases list column to long form (internal)
}
\keyword{internal}
