% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{analyze_breaking_changes}
\alias{analyze_breaking_changes}
\title{Analyze breaking change impact}
\usage{
analyze_breaking_changes(deprecated_functions, function_analysis)
}
\arguments{
\item{deprecated_functions}{Functions marked for deprecation}

\item{function_analysis}{Function analysis results}
}
\value{
Breaking change analysis
}
\description{
Analyze breaking change impact
}
