% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{analyze_function}
\alias{analyze_function}
\title{Analyze individual function}
\usage{
analyze_function(function_name)
}
\arguments{
\item{function_name}{Name of function to analyze}
}
\value{
Function analysis results
}
\description{
Analyze individual function
}
