% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{analyze_function_complexity}
\alias{analyze_function_complexity}
\title{Analyze function complexity}
\usage{
analyze_function_complexity(function_name)
}
\arguments{
\item{function_name}{Name of function}
}
\value{
Complexity analysis
}
\description{
Analyze function complexity
}
