% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{analyze_function_dependencies}
\alias{analyze_function_dependencies}
\title{Analyze function dependencies for multiple functions}
\usage{
analyze_function_dependencies(function_names)
}
\arguments{
\item{function_names}{Vector of function names to analyze}
}
\value{
List of dependencies for each function
}
\description{
Analyze function dependencies for multiple functions
}
