% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{analyze_function_usage}
\alias{analyze_function_usage}
\title{Analyze function usage patterns}
\usage{
analyze_function_usage(function_name)
}
\arguments{
\item{function_name}{Name of function}
}
\value{
Usage analysis
}
\description{
Analyze function usage patterns
}
