% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{analyze_name_coverage}
\alias{analyze_name_coverage}
\title{Analyze name coverage}
\usage{
analyze_name_coverage(transcript_data)
}
\description{
Analyze name coverage
}
\keyword{internal}
