% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{analyze_roster_attendance}
\alias{analyze_roster_attendance}
\title{Analyze roster attendance}
\usage{
analyze_roster_attendance(batch_results)
}
\description{
Analyze roster attendance
}
\keyword{internal}
