% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{analyze_timing_patterns}
\alias{analyze_timing_patterns}
\title{Analyze timing patterns}
\usage{
analyze_timing_patterns(transcript_data)
}
\description{
Analyze timing patterns
}
\keyword{internal}
