% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_name_matching_workflow.R
\name{apply_name_matching}
\alias{apply_name_matching}
\title{Apply Name Matching}
\usage{
apply_name_matching(transcript_data, name_lookup, roster_data)
}
\description{
Internal function to apply name matching to transcript data.
}
