% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{calculate_overall_quality_score}
\alias{calculate_overall_quality_score}
\title{Calculate overall quality score}
\usage{
calculate_overall_quality_score(validation_details)
}
\description{
Calculate overall quality score
}
\keyword{internal}
