% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_metrics.R
\name{calculate_overall_status}
\alias{calculate_overall_status}
\title{Calculate overall success status}
\usage{
calculate_overall_status(...)
}
\arguments{
\item{...}{All metric categories}
}
\value{
Overall success status
}
\description{
Calculate overall success status
}
