% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_audit_system.R
\name{categorize_functions}
\alias{categorize_functions}
\title{Categorize Functions by Type and Purpose}
\usage{
categorize_functions(inventory)
}
\arguments{
\item{inventory}{Function inventory from create_function_inventory()}
}
\value{
Categorized function list
}
\description{
Categorize Functions by Type and Purpose
}
\keyword{internal}
