% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{clcltcntntqltymtrcs}
\alias{clcltcntntqltymtrcs}
\title{Calculate content quality metrics}
\usage{
clcltcntntqltymtrcs(transcript_data)
}
\description{
Calculate content quality metrics
}
\keyword{internal}
