% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{clcltvrllqltyscr}
\alias{clcltvrllqltyscr}
\title{Calculate overall quality score}
\usage{
clcltvrllqltyscr(validation_details)
}
\description{
Calculate overall quality score
}
\keyword{internal}
