% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_analysis.R
\name{count_file_lines}
\alias{count_file_lines}
\title{Count lines in file}
\usage{
count_file_lines(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Number of lines
}
\description{
Count lines in file
}
