% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation_system.R
\name{create_deprecation_documentation}
\alias{create_deprecation_documentation}
\title{Create deprecation documentation}
\usage{
create_deprecation_documentation(
  deprecated_functions,
  migration_recommendations
)
}
\arguments{
\item{deprecated_functions}{Functions marked for deprecation}

\item{migration_recommendations}{Migration recommendations}
}
\value{
Deprecation documentation
}
\description{
Create deprecation documentation
}
