% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation_system.R
\name{create_deprecation_warnings}
\alias{create_deprecation_warnings}
\title{Create deprecation warnings for functions}
\usage{
create_deprecation_warnings(deprecated_functions, migration_recommendations)
}
\arguments{
\item{deprecated_functions}{Functions to deprecate}

\item{migration_recommendations}{Migration recommendations}
}
\value{
Deprecation warnings
}
\description{
Create deprecation warnings for functions
}
