% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{create_ux_integration_report}
\alias{create_ux_integration_report}
\title{Create UX integration report}
\usage{
create_ux_integration_report(
  function_categories,
  cran_functions,
  ux_categories,
  help_system
)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{cran_functions}{Functions selected for CRAN}

\item{ux_categories}{UX categories}

\item{help_system}{Help system}
}
\value{
UX integration report
}
\description{
Create UX integration report
}
