% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching_internal.R
\name{derive_speaker_column}
\alias{derive_speaker_column}
\title{Derive a 'speaker' column from common alternatives}
\usage{
derive_speaker_column(df)
}
\description{
Derive a 'speaker' column from common alternatives
}
\keyword{internal}
