% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{determine_function_category}
\alias{determine_function_category}
\title{Determine function category based on name and analysis}
\usage{
determine_function_category(func)
}
\arguments{
\item{func}{Function analysis result}
}
\value{
Category name
}
\description{
Determine function category based on name and analysis
}
