% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{determine_impact_level}
\alias{determine_impact_level}
\title{Determine impact level of deprecating a function}
\usage{
determine_impact_level(func_info)
}
\arguments{
\item{func_info}{Function analysis information}
}
\value{
Impact level
}
\description{
Determine impact level of deprecating a function
}
