% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{determine_validation_status}
\alias{determine_validation_status}
\title{Determine validation status from issues and warnings}
\usage{
determine_validation_status(validation_details)
}
\description{
Determine validation status from issues and warnings
}
\keyword{internal}
