% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_name_matching.R
\name{extract_mapped_names}
\alias{extract_mapped_names}
\title{Extract Mapped Names}
\usage{
extract_mapped_names(name_mappings)
}
\description{
Internal function to extract names from name mappings.
}
