% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_name_matching.R
\name{extract_transcript_names}
\alias{extract_transcript_names}
\title{Extract Transcript Names}
\usage{
extract_transcript_names(transcript_data)
}
\description{
Internal function to extract names from transcript data.
}
