% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{find_replacement_function}
\alias{find_replacement_function}
\title{Find replacement function for deprecated function}
\usage{
find_replacement_function(deprecated_func, cran_functions, function_analysis)
}
\arguments{
\item{deprecated_func}{Deprecated function name}

\item{cran_functions}{Available CRAN functions}

\item{function_analysis}{Function analysis results}
}
\value{
Suggested replacement function
}
\description{
Find replacement function for deprecated function
}
