% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_name_matching_workflow.R
\name{find_roster_match}
\alias{find_roster_match}
\title{Find Roster Match}
\usage{
find_roster_match(transcript_name, roster_names)
}
\description{
Internal function to find a matching name in the roster.
}
