% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{gen_detailed_validation_report}
\alias{gen_detailed_validation_report}
\title{Generate detailed validation report}
\usage{
gen_detailed_validation_report(results)
}
\description{
Generate detailed validation report
}
\keyword{internal}
