% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{gen_migration_recommendations}
\alias{gen_migration_recommendations}
\title{Generate migration recommendations}
\usage{
gen_migration_recommendations(
  deprecated_functions,
  cran_functions,
  function_analysis
)
}
\arguments{
\item{deprecated_functions}{Functions marked for deprecation}

\item{cran_functions}{Functions selected for CRAN}

\item{function_analysis}{Function analysis results}
}
\value{
Migration recommendations
}
\description{
Generate migration recommendations
}
