% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{gen_structure_recommendations}
\alias{gen_structure_recommendations}
\title{Generate structure recommendations}
\usage{
gen_structure_recommendations(results)
}
\description{
Generate structure recommendations
}
\keyword{internal}
