% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{gen_timing_recommendations}
\alias{gen_timing_recommendations}
\title{Generate timing recommendations}
\usage{
gen_timing_recommendations(results)
}
\description{
Generate timing recommendations
}
\keyword{internal}
