% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-ideal-transcripts.R
\name{generate_benchmark_summary}
\alias{generate_benchmark_summary}
\title{Generate benchmark summary statistics}
\usage{
generate_benchmark_summary(benchmark_results)
}
\description{
Generate benchmark summary statistics
}
\keyword{internal}
