% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{generate_comprehensive_summary}
\alias{generate_comprehensive_summary}
\title{Generate comprehensive summary}
\usage{
generate_comprehensive_summary(results)
}
\description{
Generate comprehensive summary
}
\keyword{internal}
