% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{generate_content_recommendations}
\alias{generate_content_recommendations}
\title{Generate content recommendations}
\usage{
generate_content_recommendations(results)
}
\description{
Generate content recommendations
}
\keyword{internal}
