% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_validation.R
\name{generate_coverage_improvement_report}
\alias{generate_coverage_improvement_report}
\title{Generate coverage improvement report}
\usage{
generate_coverage_improvement_report()
}
\value{
Coverage improvement report
}
\description{
Generate coverage improvement report
}
