% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{generate_data_quality_report}
\alias{generate_data_quality_report}
\title{Generate data quality report}
\usage{
generate_data_quality_report(session_data, summary_metrics)
}
\description{
Generate data quality report
}
\keyword{internal}
