% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-export.R
\name{generate_export_metadata}
\alias{generate_export_metadata}
\title{Generate Export Metadata}
\usage{
generate_export_metadata(data, format = "csv")
}
\description{
Generate Export Metadata
}
\keyword{internal}
