% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_audit_system.R
\name{generate_function_audit_report}
\alias{generate_function_audit_report}
\title{Generate Function Audit Report}
\usage{
generate_function_audit_report(inventory, categorization, dependencies)
}
\arguments{
\item{inventory}{Function inventory}

\item{categorization}{Function categorization}

\item{dependencies}{Function dependencies}
}
\value{
Formatted audit report
}
\description{
Generate Function Audit Report
}
\keyword{internal}
