% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{generate_function_description}
\alias{generate_function_description}
\title{Generate function description based on UX level}
\usage{
generate_function_description(func_name, ux_level)
}
\arguments{
\item{func_name}{Function name}

\item{ux_level}{UX level}
}
\value{
Function description
}
\description{
Generate function description based on UX level
}
