% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{generate_ideal_validation_recommendations}
\alias{generate_ideal_validation_recommendations}
\title{Generate ideal validation recommendations}
\usage{
generate_ideal_validation_recommendations(rule_results)
}
\description{
Generate ideal validation recommendations
}
\keyword{internal}
