% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation_system.R
\name{generate_migration_guide}
\alias{generate_migration_guide}
\title{Generate migration guide for deprecated functions}
\usage{
generate_migration_guide(deprecated_functions, migration_recommendations)
}
\arguments{
\item{deprecated_functions}{Functions marked for deprecation}

\item{migration_recommendations}{Migration recommendations}
}
\value{
Migration guide content
}
\description{
Generate migration guide for deprecated functions
}
