% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{generate_name_coverage_recs}
\alias{generate_name_coverage_recs}
\title{Generate name coverage recommendations}
\usage{
generate_name_coverage_recs(results)
}
\description{
Generate name coverage recommendations
}
\keyword{internal}
