% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{generate_progressive_disclosure_config}
\alias{generate_progressive_disclosure_config}
\title{Generate progressive disclosure configuration}
\usage{
generate_progressive_disclosure_config(ux_categories)
}
\arguments{
\item{ux_categories}{UX categories}
}
\value{
Progressive disclosure configuration
}
\description{
Generate progressive disclosure configuration
}
