% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_reduction_tracker.R
\name{generate_scope_reduction_report}
\alias{generate_scope_reduction_report}
\title{Generate Scope Reduction Progress Report}
\usage{
generate_scope_reduction_report(tracker)
}
\arguments{
\item{tracker}{Current tracker object}
}
\value{
Formatted progress report
}
\description{
Generate Scope Reduction Progress Report
}
\keyword{internal}
