% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-export.R
\name{generate_transcript_summary}
\alias{generate_transcript_summary}
\title{Generate Transcript Summary}
\usage{
generate_transcript_summary(data)
}
\description{
Generate Transcript Summary
}
\keyword{internal}
