% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{generate_validation_recs}
\alias{generate_validation_recs}
\title{Generate ideal validation recommendations}
\usage{
generate_validation_recs(rule_results)
}
\description{
Generate ideal validation recommendations
}
\keyword{internal}
