% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{get_category_allocation}
\alias{get_category_allocation}
\title{Get category allocation based on priority and remaining slots}
\usage{
get_category_allocation(category, max_functions, current_count)
}
\arguments{
\item{category}{Category name}

\item{max_functions}{Maximum total functions}

\item{current_count}{Current number of selected functions}
}
\value{
Number of functions to allocate to this category
}
\description{
Get category allocation based on priority and remaining slots
}
