% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{get_category_summary}
\alias{get_category_summary}
\title{Get category summary}
\usage{
get_category_summary(categories)
}
\arguments{
\item{categories}{Function categories}
}
\value{
Category summary
}
\description{
Get category summary
}
