% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{get_cran_functions}
\alias{get_cran_functions}
\title{Get functions for CRAN submission}
\usage{
get_cran_functions(categories, max_functions = 30)
}
\arguments{
\item{categories}{Function categories}

\item{max_functions}{Maximum number of functions for CRAN}
}
\value{
Functions selected for CRAN
}
\description{
Get functions for CRAN submission
}
