% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_metrics.R
\name{get_current_baseline}
\alias{get_current_baseline}
\title{Get Current Baseline Measurements}
\usage{
get_current_baseline()
}
\value{
List of current baseline measurements for all success metrics
}
\description{
Get Current Baseline Measurements
}
\keyword{internal}
