% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R,
%   R/scope_reduction_implementation.R
\name{get_deprecated_functions}
\alias{get_deprecated_functions}
\title{Get functions marked for deprecation}
\usage{
get_deprecated_functions()

get_deprecated_functions()
}
\arguments{
\item{categories}{Function categories}

\item{cran_functions}{Functions selected for CRAN}
}
\value{
Functions marked for deprecation

Character vector of deprecated function names
}
\description{
Get functions marked for deprecation

Get Deprecated Functions List
}
