% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_reduction_implementation.R
\name{get_essential_functions}
\alias{get_essential_functions}
\title{Get Essential Functions List}
\usage{
get_essential_functions()
}
\value{
Character vector of essential function names
}
\description{
Get Essential Functions List
}
