% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{get_function_category_by_name}
\alias{get_function_category_by_name}
\title{Get function category by name}
\usage{
get_function_category_by_name(func_name, function_categories)
}
\arguments{
\item{func_name}{Function name}

\item{function_categories}{Function categories}
}
\value{
Category name
}
\description{
Get function category by name
}
