% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{get_function_dependencies}
\alias{get_function_dependencies}
\title{Get function dependencies}
\usage{
get_function_dependencies(function_name)
}
\arguments{
\item{function_name}{Name of function}
}
\value{
Function dependencies
}
\description{
Get function dependencies
}
