% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{get_function_documentation}
\alias{get_function_documentation}
\title{Get function documentation}
\usage{
get_function_documentation(function_name)
}
\arguments{
\item{function_name}{Name of function}
}
\value{
Documentation status
}
\description{
Get function documentation
}
