% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{get_function_signature}
\alias{get_function_signature}
\title{Get function signature}
\usage{
get_function_signature(function_name)
}
\arguments{
\item{function_name}{Name of function}
}
\value{
Function signature
}
\description{
Get function signature
}
