% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{get_functions_by_category}
\alias{get_functions_by_category}
\title{Get functions by category}
\usage{
get_functions_by_category(categories, category_name)
}
\arguments{
\item{categories}{Function categories}

\item{category_name}{Name of category}
}
\value{
Functions in category
}
\description{
Get functions by category
}
