% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_reduction_implementation.R
\name{get_internal_functions}
\alias{get_internal_functions}
\title{Get Internal Functions List}
\usage{
get_internal_functions()
}
\value{
Character vector of internal function names
}
\description{
Get Internal Functions List
}
