% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{get_migration_strategy}
\alias{get_migration_strategy}
\title{Get migration strategy for deprecated function}
\usage{
get_migration_strategy(deprecated_func, replacement)
}
\arguments{
\item{deprecated_func}{Deprecated function name}

\item{replacement}{Suggested replacement}
}
\value{
Migration strategy
}
\description{
Get migration strategy for deprecated function
}
