% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_visibility_system.R
\name{get_next_level}
\alias{get_next_level}
\title{Get next UX level}
\usage{
get_next_level(current_level)
}
\arguments{
\item{current_level}{Current UX level}
}
\value{
Next UX level
}
\description{
Get next UX level
}
