% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success_metrics.R
\name{get_target_state}
\alias{get_target_state}
\title{Get Target State Definitions}
\usage{
get_target_state()
}
\value{
List of target state definitions for all success metrics
}
\description{
Get Target State Definitions
}
\keyword{internal}
