% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_name_matching_workflow.R
\name{handle_unmatched_names}
\alias{handle_unmatched_names}
\title{Handle Unmatched Names}
\usage{
handle_unmatched_names(
  unmatched_names,
  unmatched_names_action,
  privacy_level,
  data_folder,
  section_names_lookup_file
)
}
\description{
Internal function to handle unmatched names according to configuration.
}
