% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_transcripts_list.R
\name{join_transcripts_list}
\alias{join_transcripts_list}
\title{Join Transcripts Files Into a Single Tibble}
\usage{
join_transcripts_list(
  df_zoom_recorded_sessions = NULL,
  df_transcript_files = NULL,
  df_cancelled_classes = NULL
)
}
\description{
This function creates a tibble from the joining of the listing of session recordings loaded from the cloud recording
csvs
(\code{df_zoom_recorded_sessions}), the list of transcript files
(\code{df_transcript_files}), and the list of cancelled classes
(\code{df_cancelled_classes}) into a single tibble
}
\details{
loaded from the cloud recording csvs.
  zoom recordings loaded from the cloud recording csvs and transcripts.
  scheduled classes where a zoom recording is not expected.

  files or placeholders for cancelled classes.

# zoom_recorded_sessions_df <- load_zoom_recorded_sessions_list()
# transcript_files_df <- load_transcript_files_list()
# cancelled_classes_df <- load_cancelled_classes()
#
# join_transcripts_list(
#   df_zoom_recorded_sessions = zoom_recorded_sessions_df,
#   df_transcript_files = transcript_files_df,
#   df_cancelled_classes = cancelled_classes_df
# )
}
