% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{load_audit_report}
\alias{load_audit_report}
\title{Load audit report from file}
\usage{
load_audit_report(filename = "function_audit_report.rds")
}
\arguments{
\item{filename}{Input filename}
}
\value{
Audit report object
}
\description{
Load audit report from file
}
