% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cancelled_classes.R
\name{load_cancelled_classes}
\alias{load_cancelled_classes}
\title{Load Cancelled Classes csv file}
\usage{
load_cancelled_classes(
  data_folder = ".",
  cancelled_classes_file = "cancelled_classes.csv",
  cancelled_classes_col_types = "ccccccccnnnncTTcTTccci",
  write_blank_cancelled_classes = FALSE
)
}
\description{
This function creates a tibble from a provided csv file of cancelled class
sessions for scheduled classes where a zoom recording is not expected.
}
\details{
to 'data'
Defaults to 'cancelled_classes.csv'
classes. Defaults to 'ccccccccnnnncTTcTTccci'
creates a blank cancelled classes file. Defaults to FALSE

where a zoom recording is not expected.

load_cancelled_classes()
}
