% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_section_names_lookup.R
\name{load_section_names_lookup}
\alias{load_section_names_lookup}
\title{Load Section Names Lookup File}
\usage{
load_section_names_lookup(
  data_folder = ".",
  names_lookup_file = "section_names_lookup.csv",
  section_names_lookup_col_types = "ccccccccc"
)
}
\description{
This function creates a tibble from a provided file of customized student names by section.
If the file does not exist, the function prints an error and creates an empty tibble using
\code{make_blank_section_names_lookup_csv()}.
}
\details{
Defaults to 'section_names_lookup.csv'
to 'cccccccc'

load_section_names_lookup()
}
