% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_transcript_files_list.R
\name{load_transcript_files_list}
\alias{load_transcript_files_list}
\title{Load Zoom Recording Transcript Files List}
\usage{
load_transcript_files_list(
  data_folder = ".",
  transcripts_folder = "transcripts",
  transcript_files_names_pattern = "GMT\\\\d{8}-\\\\d{6}_Recording",
  dt_extract_pattern = "(?<=GMT)\\\\d{8}",
  trnscrptflxtnsnpttrn = ".transcript",
  clsdcptnflxtnsnpttrn = ".cc",
  recording_start_pattern = "(?<=GMT)\\\\d{8}-\\\\d{6}",
  recording_start_format = "\%Y\%m\%d-\%H\%M\%S",
  start_time_local_tzone = "America/Los_Angeles"
)
}
\description{
This function creates a tibble from a provided folder including
transcript files of Zoom recordings.
}
\details{
\subsection{Download Transcripts}{
\enumerate{
\item Go to \url{https://www.zoom.us/recording}
\item Click on each individual record to go to the page for that recording
\item Download the Audio Transcript and Chat File for each
\itemize{
\item Chat: \verb{GMT\\\\d\{8\}-\\\\d\{6\}_Recording.cc.vtt}
\item Transcript: \verb{GMT\\\\d\{8\}-\\\\d\{6\}_Recording.transcript.vtt}
}
\item Copy the Audio Transcript and Chat Files to \verb{data/transcripts/}
(or whatever path you identify in the \code{data_folder} and
\code{transcripts_folder} parameters).
}

to \code{data}
will store the cloud recording csvs and transcripts
transcript file names. Defaults to \verb{GMT\\\\d\{8\}-\\\\d\{6\}_Recording}
transcript from the transcript file name. Defaults to \verb{(?<=GMT)\\\\d\{8\}}
transcript files (as opposed to chat or closed caption files). Defaults to
\code{.transcript}
closed caption files (as opposed to chat or transcript files). Defaults to
\code{.cc}
start time of the transcript from the transcript file name. Defaults to
\verb{(?<=GMT)\\\\d\{8\}-\\\\d\{6\}}
recording start time of the transcript. Defaults to \verb{\\\%Y\\\%m\\\%d-\\\%H\\\%M\\\%S}
the transcript. Defaults to \code{America/Los_Angeles}

loaded from the cloud recording csvs and transcripts.

load_transcript_files_list()
}
}
