% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zoom_recorded_sessions_list.R
\name{load_zoom_recorded_sessions_list}
\alias{load_zoom_recorded_sessions_list}
\title{Load Zoom Recorded Sessions List}
\usage{
load_zoom_recorded_sessions_list(
  data_folder = ".",
  transcripts_folder = "transcripts",
  topic_split_pattern = paste0("^(?<dept>\\\\S+) (?<course_section>\\\\S+) - ",
    "(?<day>[A-Za-z]+) (?<time>\\\\S+\\\\s*\\\\S+) (?<instructor>\\\\(.*?\\\\))"),
  zmrcrddsssnscsvnmspttrn =
    "zoomus_recordings__\\\\d{8}(?:\\\\s+copy\\\\s*\\\\d*)?\\\\.csv",
  zmrcrddsssnscsvclnms = paste("Topic", "ID", "Start Time", "File Size (MB)",
    "File Count", "Total Views", "Total Downloads", "Last Accessed", sep = ","),
  dept = "LTF",
  semester_start_mdy = "Jan 01, 2024",
  scheduled_session_length_hours = 1.5,
  verbose = FALSE
)
}
\description{
This function creates a tibble from a provided csv file of Zoom recordings.
}
\details{
\subsection{Download Zoom csv file with list of recordings and transcripts}{
\enumerate{
\item Go to \url{https://www.zoom.us/recording}
\item Export the Cloud Recordings
\item Copy the cloud recording csv (naming convention:
\verb{zoomus_recordings__\\\\d\{8\}\\.csv}) to \verb{data/transcripts/}
(or whatever path you identify in the \code{data_folder} and
\code{transcripts_folder} parameters).
}
}
}
