% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zoom_transcript.R
\name{load_zoom_transcript}
\alias{load_zoom_transcript}
\title{Load Zoom Transcript}
\usage{
load_zoom_transcript(transcript_file_path = NULL)
}
\arguments{
\item{transcript_file_path}{Path to the transcript file to load}
}
\value{
A tibble containing the transcript data, or NULL if the file is empty
}
\description{
Load a Zoom recording transcript and return tibble containing the comments from a Zoom recording transcript
}
\details{
Original code posted by Conor Healy:
https://ucbischool.slack.com/archives/C02A36407K9/p1631855705002000 Addition
of \code{wordcount} by Brooks Ambrose:
https://gist.github.com/brooksambrose/1a8a673eb3bf884c1868ad4d80f08246
}
\examples{
# Load a sample transcript from the package's extdata directory
transcript_file <- system.file("extdata/test_transcripts/intro_statistics_week1.vtt",
  package = "engager"
)
load_zoom_transcript(transcript_file_path = transcript_file)

}
