% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clean_names_df.R
\name{make_clean_names_df}
\alias{make_clean_names_df}
\title{Make Clean Names DF}
\usage{
make_clean_names_df(
  data_folder = ".",
  section_names_lookup_file = "section_names_lookup.csv",
  transcripts_metrics_df = NULL,
  roster_sessions = NULL
)
}
\arguments{
\item{data_folder}{overall data folder for your recordings. Defaults to
'data'}

\item{section_names_lookup_file}{File name of the csv file of customized
student names by section Defaults to 'section_names_lookup.csv'}

\item{transcripts_metrics_df}{A tibble containing session details and summary
metrics by speaker for all class sessions in the tibble provided.}

\item{roster_sessions}{A tibble listing the students enrolled in the class or
classes, with rows for each recorded class section for each student.}
}
\value{
A tibble containing session details and summary metrics by speaker
for all class sessions (and placeholders for missing sections), including
customized student names.
}
\description{
This function creates a tibble containing session details and summary metrics
by speaker for all class sessions (and placeholders for missing sections)
from the joining of:
\itemize{
\item a tibble of customized student names by section (\code{section_names_lookup_file} in the \code{data_folder} folder),
\item a tibble containing session details and summary metrics by speaker for all class sessions
(\code{transcripts_metrics_df}), and
\item a tibble listing the students enrolled in the class or classes, with rows for each recorded class section for each
student (\code{roster_sessions}) into a single tibble.
}
}
\keyword{"101"),}
\keyword{"101.B")}
\keyword{"101.B"),}
\keyword{"11:00"),}
\keyword{"2024-01-02}
\keyword{"67890"),}
\keyword{"B"),}
\keyword{"CS"),}
\keyword{"Excellent"),}
\keyword{"Jane}
\keyword{"Tuesday"),}
\keyword{"data",}
\keyword{"section_names_lookup.csv",}
\keyword{)}
\keyword{0.15),}
\keyword{1)}
\keyword{1),}
\keyword{100),}
\keyword{101),}
\keyword{10:00:00",}
\keyword{11:00:00"),}
\keyword{15),}
\keyword{450),}
\keyword{750),}
\keyword{<-}
\keyword{=}
\keyword{Doe"),}
\keyword{Smith",}
\keyword{c("101",}
\keyword{c("101.A",}
\keyword{c("10:00",}
\keyword{c("12345",}
\keyword{c("2024-01-01}
\keyword{c("A",}
\keyword{c("CS",}
\keyword{c("Good",}
\keyword{c("John}
\keyword{c("Monday",}
\keyword{c(0.1,}
\keyword{c(1,}
\keyword{c(10,}
\keyword{c(100,}
\keyword{c(101,}
\keyword{c(300,}
\keyword{c(500,}
\keyword{comments}
\keyword{course}
\keyword{course_section}
\keyword{data_folder}
\keyword{day}
\keyword{dept}
\keyword{duration}
\keyword{duration_perc}
\keyword{first_last}
\keyword{internal}
\keyword{make_clean_names_df(}
\keyword{n}
\keyword{n_perc}
\keyword{name}
\keyword{name_raw}
\keyword{preferred_name}
\keyword{roster_sessions}
\keyword{sample_roster}
\keyword{sample_transcripts,}
\keyword{section}
\keyword{section_names_lookup_file}
\keyword{session_num}
\keyword{start_time_local}
\keyword{student_id}
\keyword{tibble::tibble(}
\keyword{time}
\keyword{transcripts_metrics_df}
\keyword{wordcount}
\keyword{wordcount_perc}
\keyword{wpm}
