% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roster_small.R
\name{make_roster_small}
\alias{make_roster_small}
\title{Make a Smaller DF of the Student Roster}
\usage{
make_roster_small(roster_df = NULL)
}
\arguments{
\item{roster_df}{A tibble listing the students enrolled in the class or
classes with a small subset of the roster columns. Must contain the following columns:
\itemize{
\item student_id: character
\item first_last: character
\item preferred_name: character
\item dept: character
\item course: character
\item section: character
}}
}
\value{
A tibble listing the students enrolled in the class or classes with a
small subset of the roster columns.
}
\description{
This function creates a tibble that includes rows for each students enrolled
in the class or classes, with a small subset of the roster columns.
}
\keyword{deprecated}
\keyword{internal}
\keyword{roster}
