% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sections_df.R
\name{make_sections_df}
\alias{make_sections_df}
\title{Make a DF of Class Sections from the Student Roster}
\usage{
make_sections_df(roster_df = NULL)
}
\arguments{
\item{roster_df}{A tibble listing the students enrolled in the class or classes.
Must contain the following columns:
\itemize{
\item dept: character
\item course: character
\item section: character
}}
}
\value{
A tibble with the following columns:
\itemize{
\item dept: character
\item course: character
\item section: character
\item n: integer (count of students in each section)
}
}
\description{
This function creates a tibble that includes rows for each section (grouped by dept and course number) and student
count in each.
}
\keyword{=}
\keyword{deprecated}
\keyword{internal}
\keyword{make_sections_df(roster_df}
\keyword{roster_df)}
\keyword{sections}
