% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_student_roster_sessions.R
\name{make_student_roster_sessions}
\alias{make_student_roster_sessions}
\title{Make a DF of the Student Roster With Rows for Each Recorded Class Section}
\usage{
make_student_roster_sessions(
  transcripts_list_df = NULL,
  roster_small_df = NULL
)
}
\arguments{
\item{transcripts_list_df}{A tibble listing the class sessions with corresponding transcript
files or placeholders for cancelled classes.}

\item{roster_small_df}{A tibble listing the students enrolled in the class or classes with a
small subset of the roster columns.}
}
\value{
A tibble listing the students enrolled in the class or classes, with rows for each recorded class section for
each student.
}
\description{
This function creates a tibble from a provided tibble students enrolled
in the class or classes (\code{roster_small_df}) and a tibble of class sessions with corresponding transcript
files or placeholders for cancelled classes (\code{transcripts_list_df}).
}
\keyword{)}
\keyword{),}
\keyword{=}
\keyword{df_cancelled_classes}
\keyword{df_transcript_files}
\keyword{df_zoom_recorded_sessions}
\keyword{internal}
\keyword{join_transcripts_list(}
\keyword{load_cancelled_classes()}
\keyword{load_transcript_files_list(),}
\keyword{load_zoom_recorded_sessions_list(),}
\keyword{make_roster_small(roster_df}
\keyword{make_student_roster_sessions(}
\keyword{roster_df)}
\keyword{roster_small_df}
\keyword{transcripts_list_df}
