% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_students_only_transcripts_summary_df.R
\name{make_students_only_transcripts_summary_df}
\alias{make_students_only_transcripts_summary_df}
\title{Make Students Only Transcripts Summary}
\usage{
make_students_only_transcripts_summary_df(
  transcripts_session_summary_df = NULL,
  preferred_name_exclude_cv = c("dead_air", "Instructor Name", "Guests", "unknown")
)
}
\description{
This function creates a tibble from summary results at the
level of the class section and preferred student name after filtering for only the students enrolled in the class.
}
