% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_audit_system.R
\name{map_function_dependencies}
\alias{map_function_dependencies}
\title{Map Function Dependencies}
\usage{
map_function_dependencies(inventory)
}
\arguments{
\item{inventory}{Function inventory from create_function_inventory()}
}
\value{
Dependency map
}
\description{
Map Function Dependencies
}
\keyword{internal}
