% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{mark_deprecated_functions}
\alias{mark_deprecated_functions}
\title{Mark functions for deprecation}
\usage{
mark_deprecated_functions(function_categories, cran_functions)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{cran_functions}{Functions selected for CRAN}
}
\value{
Functions marked for deprecation
}
\description{
Mark functions for deprecation
}
