% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-ideal-transcripts.R
\name{measure_memory_usage}
\alias{measure_memory_usage}
\title{Measure memory usage of a function}
\usage{
measure_memory_usage(expr)
}
\description{
Measure memory usage of a function
}
\keyword{internal}
