% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecation_system.R
\name{pdtnmspcfrdprctn}
\alias{pdtnmspcfrdprctn}
\title{Update NAMESPACE for deprecation}
\usage{
pdtnmspcfrdprctn(cran_functions, deprecated_functions)
}
\arguments{
\item{cran_functions}{Functions to keep in NAMESPACE}

\item{deprecated_functions}{Functions to remove from NAMESPACE}
}
\value{
Updated NAMESPACE content
}
\description{
Update NAMESPACE for deprecation
}
