% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching_internal.R
\name{prepare_transcript_names}
\alias{prepare_transcript_names}
\title{Prepare transcript names: normalize and hash speaker names
Expects column 'speaker'. Optional 'timestamp'.}
\usage{
prepare_transcript_names(transcripts_df, key = NULL)
}
\description{
Prepare transcript names: normalize and hash speaker names
Expects column 'speaker'. Optional 'timestamp'.
}
\keyword{internal}
