% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{prepare_visualization_data}
\alias{prepare_visualization_data}
\title{Prepare visualization data}
\usage{
prepare_visualization_data(comparison_data, metrics)
}
\description{
Prepare visualization data
}
\keyword{internal}
