% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-ideal-transcripts.R
\name{print_benchmark_summary}
\alias{print_benchmark_summary}
\title{Print benchmark summary}
\usage{
print_benchmark_summary(summary)
}
\description{
Print benchmark summary
}
\keyword{internal}
