% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{print_category_summary}
\alias{print_category_summary}
\title{Print category summary}
\usage{
print_category_summary(categories)
}
\arguments{
\item{categories}{Function categories}
}
\description{
Print category summary
}
