% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{print_ux_category_summary}
\alias{print_ux_category_summary}
\title{Print UX category summary}
\usage{
print_ux_category_summary(ux_categories)
}
\arguments{
\item{ux_categories}{UX categories}
}
\description{
Print UX category summary
}
