% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{print_validation_summary}
\alias{print_validation_summary}
\title{Print validation summary}
\usage{
print_validation_summary(validation_results)
}
\arguments{
\item{validation_results}{Validation results}
}
\description{
Print validation summary
}
