% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{prioritize_functions_for_ux}
\alias{prioritize_functions_for_ux}
\title{Prioritize functions for UX levels}
\usage{
prioritize_functions_for_ux(function_categories, cran_functions)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{cran_functions}{Functions selected for CRAN}
}
\value{
Prioritized function list
}
\description{
Prioritize functions for UX levels
}
