% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_analysis.R
\name{rank_coverage_impact}
\alias{rank_coverage_impact}
\title{Rank files by coverage impact}
\usage{
rank_coverage_impact(low_coverage)
}
\arguments{
\item{low_coverage}{Files with low coverage}
}
\value{
Impact ranking
}
\description{
Rank files by coverage impact
}
