% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_matching_internal.R
\name{resolve_name_hash_key}
\alias{resolve_name_hash_key}
\title{Resolve the name hash key using precedence}
\usage{
resolve_name_hash_key(key)
}
\description{
Precedence: explicit function argument >
Sys.getenv("ENGAGER_NAME_HASH_KEY") >
getOption("engager.name_hash_key", NULL)
}
\keyword{internal}
