% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_function_audit.R
\name{save_audit_report}
\alias{save_audit_report}
\title{Save audit report to file}
\usage{
save_audit_report(audit_report, filename = "function_audit_report.rds")
}
\arguments{
\item{audit_report}{Audit report object}

\item{filename}{Output filename}
}
\description{
Save audit report to file
}
