% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{select_cran_functions}
\alias{select_cran_functions}
\title{CRAN Optimization System}
\usage{
select_cran_functions(function_categories, max_functions = 30)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{max_functions}{Maximum number of functions for CRAN (default: 30)}
}
\value{
CRAN-optimized function set
}
\description{
Optimizes function set for CRAN submission
}
\keyword{internal}
