% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R
\name{test_categorization_system}
\alias{test_categorization_system}
\title{Test categorization system}
\usage{
test_categorization_system()
}
\value{
Test results
}
\description{
Test categorization system
}
