% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_optimization.R
\name{test_cran_optimization}
\alias{test_cran_optimization}
\title{Test CRAN optimization system}
\usage{
test_cran_optimization()
}
\value{
Test results
}
\description{
Test CRAN optimization system
}
