% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{update_function_descriptions}
\alias{update_function_descriptions}
\title{Update function help descriptions}
\usage{
update_function_descriptions(function_categories, ux_categories)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{ux_categories}{UX categories}
}
\value{
Updated function descriptions
}
\description{
Update function help descriptions
}
