% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_integration.R
\name{update_help_system}
\alias{update_help_system}
\title{Update help system with new categories}
\usage{
update_help_system(function_categories, ux_categories)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{ux_categories}{UX categories}
}
\value{
Updated help system
}
\description{
Update help system with new categories
}
