% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_reduction_tracker.R
\name{update_scope_reduction_tracker}
\alias{update_scope_reduction_tracker}
\title{Update Scope Reduction Tracker}
\usage{
update_scope_reduction_tracker(tracker, updates)
}
\arguments{
\item{tracker}{Current tracker object}

\item{updates}{List of updates to apply}
}
\value{
Updated tracker object
}
\description{
Update Scope Reduction Tracker
}
\keyword{internal}
