% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_categorization.R,
%   R/validation_system.R
\name{validate_categories}
\alias{validate_categories}
\title{Validate function categories}
\usage{
validate_categories(function_categories)

validate_categories(function_categories)
}
\arguments{
\item{function_categories}{Function categories from audit}

\item{categories}{Function categories}
}
\value{
Validation results

Category validation results
}
\description{
Validate function categories

Validate function categories
}
