% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_chronological_order}
\alias{validate_chronological_order}
\title{Validate chronological order}
\usage{
validate_chronological_order(transcript_data)
}
\description{
Validate chronological order
}
\keyword{internal}
