% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{validate_comment_content}
\alias{validate_comment_content}
\title{Validate comment content}
\usage{
validate_comment_content(session_data)
}
\description{
Validate comment content
}
\keyword{internal}
