% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_content_diversity}
\alias{validate_content_diversity}
\title{Validate content diversity}
\usage{
validate_content_diversity(transcript_data)
}
\description{
Validate content diversity
}
\keyword{internal}
