% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_dependencies}
\alias{validate_dependencies}
\title{Validate function dependencies}
\usage{
validate_dependencies(cran_functions, function_analysis)
}
\arguments{
\item{cran_functions}{Functions selected for CRAN}

\item{function_analysis}{Function analysis results}
}
\value{
Dependency validation results
}
\description{
Validate function dependencies
}
