% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_dialogue_length}
\alias{validate_dialogue_length}
\title{Validate dialogue length}
\usage{
validate_dialogue_length(transcript_data)
}
\description{
Validate dialogue length
}
\keyword{internal}
