% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_documentation}
\alias{validate_documentation}
\title{Validate function documentation}
\usage{
validate_documentation(cran_functions, function_analysis)
}
\arguments{
\item{cran_functions}{Functions selected for CRAN}

\item{function_analysis}{Function analysis results}
}
\value{
Documentation validation results
}
\description{
Validate function documentation
}
