% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_duration_calculations}
\alias{validate_duration_calculations}
\title{Validate duration calculations}
\usage{
validate_duration_calculations(transcript_data)
}
\description{
Validate duration calculations
}
\keyword{internal}
