% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{validate_engagement_metrics}
\alias{validate_engagement_metrics}
\title{Validate engagement metrics}
\usage{
validate_engagement_metrics(summary_metrics, min_comments)
}
\description{
Validate engagement metrics
}
\keyword{internal}
