% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_expected_names}
\alias{validate_expected_names}
\title{Validate expected names}
\usage{
validate_expected_names(transcript_data, expected_names)
}
\description{
Validate expected names
}
\keyword{internal}
