% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_section_names_lookup.R
\name{validate_lookup_file_format}
\alias{validate_lookup_file_format}
\title{Validate Lookup File Format}
\usage{
validate_lookup_file_format(data)
}
\arguments{
\item{data}{The data to validate}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Internal function to validate the format of section names lookup files.
}
\keyword{internal}
