% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{validate_name_consistency}
\alias{validate_name_consistency}
\title{Validate name consistency}
\usage{
validate_name_consistency(session_data)
}
\description{
Validate name consistency
}
\keyword{internal}
