% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_name_edge_cases}
\alias{validate_name_edge_cases}
\title{Validate name edge cases}
\usage{
validate_name_edge_cases(transcript_data)
}
\description{
Validate name edge cases
}
\keyword{internal}
