% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_name_variations}
\alias{validate_name_variations}
\title{Validate name variations}
\usage{
validate_name_variations(transcript_data)
}
\description{
Validate name variations
}
\keyword{internal}
