% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{validate_participant_counts}
\alias{validate_participant_counts}
\title{Validate participant counts}
\usage{
validate_participant_counts(participation_patterns, rules)
}
\description{
Validate participant counts
}
\keyword{internal}
