% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_realistic_patterns}
\alias{validate_realistic_patterns}
\title{Validate realistic patterns}
\usage{
validate_realistic_patterns(transcript_data)
}
\description{
Validate realistic patterns
}
\keyword{internal}
