% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_required_fields}
\alias{validate_required_fields}
\title{Validate required fields}
\usage{
validate_required_fields(transcript_data, strict_mode)
}
\description{
Validate required fields
}
\keyword{internal}
