% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_scenario_completeness}
\alias{validate_scenario_completeness}
\title{Validate scenario completeness}
\usage{
validate_scenario_completeness(transcript_data)
}
\description{
Validate scenario completeness
}
\keyword{internal}
