% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate that a data frame contains required columns (and optionally types)}
\usage{
validate_schema(df = NULL, required_cols = NULL, types = NULL)
}
\description{
df <- tibble::tibble(a = 1L, b = "x")
validate_schema(df, c("a", "b"), types = list(a = "integer", b = "character"))
}
