% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-transcript-validation.R
\name{validate_speaker_consistency}
\alias{validate_speaker_consistency}
\title{Validate speaker consistency}
\usage{
validate_speaker_consistency(transcript_data)
}
\description{
Validate speaker consistency
}
\keyword{internal}
