% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_system.R
\name{validate_test_coverage}
\alias{validate_test_coverage}
\title{Validate test coverage}
\usage{
validate_test_coverage(cran_functions, function_analysis)
}
\arguments{
\item{cran_functions}{Functions selected for CRAN}

\item{function_analysis}{Function analysis results}
}
\value{
Test coverage validation results
}
\description{
Validate test coverage
}
