% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_engagement_metrics.R
\name{write_engagement_metrics}
\alias{write_engagement_metrics}
\title{Write Engagement Metrics to CSV}
\usage{
write_engagement_metrics(
  metrics_data = NULL,
  file_path = NULL,
  comments_format = c("text", "count")
)
}
\description{
Deprecated: use \code{write_metrics(data, what = 'engagement', path = ...)} instead.
}
\details{
comments)
}
