% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_section_names_lookup.R
\name{write_section_names_lookup}
\alias{write_section_names_lookup}
\title{Write Section Names Lookup}
\usage{
write_section_names_lookup(
  clean_names_df = NULL,
  data_folder = ".",
  section_names_lookup_file = "section_names_lookup.csv"
)
}
\arguments{
\item{clean_names_df}{A tibble containing session details and summary metrics
by speaker for all class sessions (and placeholders for missing sections),
including customized student names.}

\item{data_folder}{Overall data folder for your recordings and data. Defaults
to 'data'}

\item{section_names_lookup_file}{File name of the csv file of customized
student names by section Defaults to 'section_names_lookup.csv'}
}
\value{
A tibble corresponding to the csv file saved, which is a sorted subset
of the provided tibble containing session details and summary metrics by
speaker for all class sessions (and placeholders for missing sections),
including customized student names.
}
\description{
This function takes a tibble containing session details and summary metrics
by speaker for all class sessions (and placeholders for missing sections),
including customized student names (\code{clean_names_df}) and saves a subset as a
csv file with the specified file name (\code{section_names_lookup_file}) to the
specified data folder (\code{data_folder}).
}
\keyword{"101"),}
\keyword{"101",}
\keyword{"101.A",}
\keyword{"101.B")}
\keyword{"101.B"),}
\keyword{"10:00}
\keyword{"2024-01-01}
\keyword{"2024-01-02}
\keyword{"9:00}
\keyword{"A",}
\keyword{"Average"),}
\keyword{"B"),}
\keyword{"B456",}
\keyword{"C789"),}
\keyword{"CS"),}
\keyword{"CS",}
\keyword{"Excellent",}
\keyword{"Monday",}
\keyword{"Student}
\keyword{"Tuesday"),}
\keyword{"section_names_lookup.csv"}
\keyword{"section_names_lookup.csv",}
\keyword{#}
\keyword{)}
\keyword{),}
\keyword{0.3,}
\keyword{0.5),}
\keyword{09:00:00",}
\keyword{1,}
\keyword{100),}
\keyword{100,}
\keyword{101),}
\keyword{101,}
\keyword{10:00:00"),}
\keyword{12),}
\keyword{2)}
\keyword{2),}
\keyword{240,}
\keyword{360),}
\keyword{400,}
\keyword{600),}
\keyword{8,}
\keyword{<-}
\keyword{=}
\keyword{A",}
\keyword{AM"),}
\keyword{AM",}
\keyword{B",}
\keyword{C"),}
\keyword{Clean}
\keyword{Create}
\keyword{Run}
\keyword{TRUE)}
\keyword{a}
\keyword{c("101",}
\keyword{c("101.A",}
\keyword{c("2024-01-01}
\keyword{c("9:00}
\keyword{c("A",}
\keyword{c("A123",}
\keyword{c("CS",}
\keyword{c("Good",}
\keyword{c("Monday",}
\keyword{c("Student}
\keyword{c(0.4,}
\keyword{c(1,}
\keyword{c(10,}
\keyword{c(100,}
\keyword{c(101,}
\keyword{c(300,}
\keyword{c(500,}
\keyword{clean_names_df}
\keyword{comments}
\keyword{course}
\keyword{course_section}
\keyword{data_folder}
\keyword{day}
\keyword{dept}
\keyword{dir.create(temp_dir)}
\keyword{directory}
\keyword{duration}
\keyword{duration_perc}
\keyword{example}
\keyword{first_last}
\keyword{for}
\keyword{internal}
\keyword{make_clean_names_df(}
\keyword{n}
\keyword{n_perc}
\keyword{name_raw}
\keyword{preferred_name}
\keyword{recursive}
\keyword{roster_sessions}
\keyword{sample_roster}
\keyword{sample_transcript_list,}
\keyword{section}
\keyword{section_names_lookup_file}
\keyword{session_num}
\keyword{start_time_local}
\keyword{student_id}
\keyword{temp_dir}
\keyword{temp_dir,}
\keyword{tempfile("example")}
\keyword{temporary}
\keyword{the}
\keyword{tibble::tibble(}
\keyword{time}
\keyword{transcripts_metrics_df}
\keyword{unlink(temp_dir,}
\keyword{up}
\keyword{with}
\keyword{wordcount}
\keyword{wordcount_perc}
\keyword{wpm}
\keyword{write_section_names_lookup(}
