% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transcripts_session_summary.R
\name{write_transcripts_session_summary}
\alias{write_transcripts_session_summary}
\title{Write Transcripts Session Summary}
\usage{
write_transcripts_session_summary(
  transcripts_session_summary_df = NULL,
  data_folder = ".",
  transcripts_session_summary_file = "transcripts_session_summary.csv"
)
}
\description{
Deprecated: use \code{write_metrics(data, what = 'session_summary', path = ...)} instead.
}
