% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_transcripts_summary.R
\name{write_transcripts_summary}
\alias{write_transcripts_summary}
\title{Write Transcripts Summary}
\usage{
write_transcripts_summary(
  transcripts_summary_df = NULL,
  data_folder = ".",
  transcripts_summary_file = "transcripts_summary.csv"
)
}
\description{
Deprecated: use \code{write_metrics(data, what = 'summary', path = ...)} instead.
}
\details{
student name.
}
