% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_audit_system.R,
%   R/function_categorization.R
\name{categorize_functions}
\alias{categorize_functions}
\title{Categorize Functions by Type and Purpose}
\usage{
categorize_functions(function_analysis)

categorize_functions(function_analysis)
}
\arguments{
\item{function_analysis}{List of function analysis results}

\item{inventory}{Function inventory from create_function_inventory()}
}
\value{
Categorized function list

Function categories
}
\description{
Categorizes functions by technical purpose and CRAN readiness
}
\keyword{internal}
