% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engager-package.R
\docType{package}
\name{engager-package}
\alias{engager}
\alias{engager-package}
\title{engager: Student Engagement Analysis Package}
\description{
Tools for processing Zoom transcripts, matching names with rosters,
computing engagement metrics, and creating privacy-aware visualizations
and reports. The package emphasizes privacy-supporting defaults,
FERPA-oriented review helpers, and masked outputs by default.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/revgizmo/engager}
  \item \url{https://revgizmo.github.io/engager/}
  \item Report bugs at \url{https://github.com/revgizmo/engager/issues}
}

}
\author{
\strong{Maintainer}: Conor Healy \email{conorhealy@berkeley.edu}

}
