% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_interactive_help.R
\name{find_function_for_task}
\alias{find_function_for_task}
\title{Interactive Help System}
\usage{
find_function_for_task(task)
}
\arguments{
\item{task}{Description of what user wants to do}
}
\description{
Provides interactive help and function discovery to help users
find the right functions for their specific tasks.

Helps users find the right function based on what they want to do
}
\examples{
\dontrun{
find_function_for_task('load transcript file')
find_function_for_task('create visualizations')
find_function_for_task('export results')
}
}
