% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideal-course-batch.R
\name{generate_validation_recommendations}
\alias{generate_validation_recommendations}
\title{Generate validation recommendations}
\usage{
generate_validation_recommendations(rule_results, validation_summary)
}
\description{
Generate validation recommendations
}
\keyword{internal}
