% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope_reduction_implementation.R
\name{get_deprecated_functions}
\alias{get_deprecated_functions}
\title{Get Deprecated Functions List}
\usage{
get_deprecated_functions()
}
\value{
Character vector of deprecated function names
}
\description{
Get Deprecated Functions List
}
