% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_visibility_system.R
\name{get_ux_level}
\alias{get_ux_level}
\title{Get current user experience level}
\usage{
get_ux_level()
}
\value{
Current UX level
}
\description{
Get current user experience level
}
\examples{
\dontrun{
current_level <- get_ux_level()
}
}
