% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferpa_compliance.R
\name{identify_anonymization_columns}
\alias{identify_anonymization_columns}
\title{Identify Columns to Anonymize}
\usage{
identify_anonymization_columns(data, preserve_columns)
}
\description{
Helper function to identify which columns in a dataset should be anonymized
based on common PII column names.
}
