% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_roster.R
\name{load_roster}
\alias{load_roster}
\title{Load a roster file}
\usage{
load_roster(
  data_folder = ".",
  roster_file = "roster.csv",
  strict_errors = FALSE
)
}
\arguments{
\item{data_folder}{Directory containing the roster file}

\item{roster_file}{Name of the roster CSV file}

\item{strict_errors}{Logical; if TRUE, throw error when file doesn't exist}
}
\value{
A tibble with roster data, filtered by enrolled status if applicable
}
\description{
Load a roster file
}
