% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sections_df.R
\name{make_sections_df}
\alias{make_sections_df}
\title{Make a DF of Class Sections from the Student Roster}
\usage{
make_sections_df(roster_df = NULL)
}
\description{
This function creates a tibble that includes rows for each section (grouped by dept and course number) and student
count in each.
Must contain the following columns:
\itemize{
\item dept: character
\item course: character
\item section: character
}
}
\details{
\itemize{
\item dept: character
\item course: character
\item section: character
\item n: integer (count of students in each section)
}

make_sections_df(roster_df = roster_df)
}
