% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_error_handling.R
\name{show_error_recovery}
\alias{show_error_recovery}
\title{Show error recovery suggestions}
\usage{
show_error_recovery(error_type)
}
\arguments{
\item{error_type}{Type of error encountered}
}
\description{
Provides specific recovery suggestions based on error type
}
\examples{
\dontrun{
show_error_recovery('file_not_found')
show_error_recovery('permission_denied')
}
}
