% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_interactive_help.R
\name{show_function_categories}
\alias{show_function_categories}
\title{Show function categories and counts}
\usage{
show_function_categories()
}
\description{
Show function categories and counts
}
\examples{
\dontrun{
show_function_categories()
}
}
