% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_guidance_system.R
\name{show_function_help}
\alias{show_function_help}
\title{Show help for specific function}
\usage{
show_function_help(function_name)
}
\arguments{
\item{function_name}{Name of function to get help for}
}
\description{
Show help for specific function
}
\examples{
\dontrun{
show_function_help("load_zoom_transcript")
show_function_help("basic_transcript_analysis")
}
}
