% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_guidance_system.R
\name{show_privacy_guidance}
\alias{show_privacy_guidance}
\title{Show privacy and ethics guidance}
\usage{
show_privacy_guidance()
}
\description{
Show privacy and ethics guidance
}
\examples{
\dontrun{
show_privacy_guidance()
}
}
