% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ux_guidance_system.R
\name{show_troubleshooting}
\alias{show_troubleshooting}
\title{Show troubleshooting guide}
\usage{
show_troubleshooting()
}
\description{
Show troubleshooting guide
}
\examples{
\dontrun{
show_troubleshooting()
}
}
